PIDE implementations may be resolved from the classpath
* resolver: abstraction over finding PIDE implementations
* coursier-based resolver is maintained
* new classpath-based resolver looks up JAR from classpath
* build: `pide-package` module bundles PIDE implementations as resources
