+++ setup/src/main/scala/Setup.scala
@@ -9,8 +9,6 @@ import scala.util._
 
 import cats.data.Xor
 
-import coursier._
-
 import org.log4s._
 
 import edu.tum.cs.isabelle.api.{BuildInfo, Environment, Version}
@@ -108,65 +106,6 @@ object Setup {
         }
     }
 
-  def fetchImplementation(platform: Platform, version: Version)(implicit ec: ExecutionContext): Future[List[Path]] = {
-    val base = platform.versionedStorage
-
-    val repositories = Seq(
-      coursier.Cache.ivy2Local,
-      MavenRepository("https://repo1.maven.org/maven2"),
-      MavenRepository("https://oss.sonatype.org/content/repositories/releases")
-    )
-
-    val downloadLogger = new coursier.Cache.Logger {
-      override def downloadingArtifact(url: String, file: File) =
-        logger.debug(s"Downloading artifact from $url ...")
-      override def downloadedArtifact(url: String, success: Boolean) = {
-        val file = url.split('/').last
-        if (success)
-          logger.debug(s"Successfully downloaded $file")
-        else
-          logger.error(s"Failed to download $file")
-      }
-    }
-
-    val fetch = Fetch.from(
-      repositories,
-      Cache.fetch(logger = Some(downloadLogger), cachePolicy = CachePolicy.LocalOnly),
-      Cache.fetch(logger = Some(downloadLogger), cachePolicy = CachePolicy.FetchMissing)
-    )
-
-    val dependency =
-      Dependency(
-        Module(BuildInfo.organization, s"pide-${version.identifier}_${BuildInfo.scalaBinaryVersion}"),
-        BuildInfo.version
-      )
-
-    def fetchArtifact(artifact: Artifact, cachePolicy: CachePolicy) =
-      Cache.file(artifact, logger = Some(downloadLogger), cachePolicy = cachePolicy)
-
-    platform.withAsyncLock { () =>
-      for {
-        resolution <- Resolution(Set(dependency)).process.run(fetch).toScalaFuture
-        artifacts = {
-            if (!resolution.isDone)
-              sys.error("not converged")
-            else if (!resolution.errors.isEmpty)
-              sys.error(s"errors: ${resolution.errors}")
-            else if (!resolution.conflicts.isEmpty)
-              sys.error(s"conflicts: ${resolution.conflicts}")
-            else
-              resolution.artifacts.toSet
-          }
-        res <- Future.traverse(artifacts.toList)(artifact =>
-          (fetchArtifact(artifact, CachePolicy.LocalOnly)
-             orElse fetchArtifact(artifact, CachePolicy.FetchMissing))
-            .run.toScalaFuture
-        )
-      }
-      yield
-        res.map(_.fold(err => sys.error(err.toString), _.toPath))
-    }
-  }
 
 }
 
@@ -206,13 +145,16 @@ final case class Setup(home: Path, platform: Platform, version: Version, package
     constructor.newInstance(context)
   }
 
+  def makeEnvironment(implicit ec: ExecutionContext): Future[Environment] =
+    makeEnvironment(Resolver.Default)
+
   /**
    * Prepares a fresh [[edu.tum.cs.isabelle.api.Environment]].
    *
-   * Uses [[Setup.fetchImplementation]] to download the required classpath. It
+   * If the [[Resolver resolver]] found an appropriate classpath, this method
    * also checks for matching [[edu.tum.cs.isabelle.api.BuildInfo build info]].
    */
-  def makeEnvironment(implicit ec: ExecutionContext): Future[Environment] =
-    Setup.fetchImplementation(platform, version).map(paths => instantiate(paths.map(_.toUri.toURL)))
+  def makeEnvironment(resolver: Resolver)(implicit ec: ExecutionContext): Future[Environment] =
+    resolver.resolve(platform, version).map(paths => instantiate(paths.map(_.toUri.toURL)))
 
 }
